refactor: started work on some cleanup
