+++ assets/templates/pages/coffees.html
@@ -6,10 +6,10 @@
   </ul>
 </div>
 
-{{ if not .Data.open }}
+{{ if not .Open }}
     <button class="btn btn-primary" onclick="this.remove(); $('#create-card').classList.remove('hidden')">Create Coffee</button>
 {{ end}}
-<div class="card card-border bg-neutral w-full {{ if not .Data.open }}hidden{{ end }}" id="create-card">
+<div class="card card-border bg-neutral w-full {{ if not .Open }}hidden{{ end }}" id="create-card">
     <div class="card-body">
         <form
             hx-post="/coffees"
@@ -26,7 +26,7 @@
                     <span class="label w-28">Roaster *</span>
                     <select class="w-full" name="roaster.int">
                         <option value="" disabled selected>Pick a Roaster</option>
-                        {{ range .Data.Roasters }}
+                        {{ range .Roasters }}
                             <option value="{{ .ID }}" {{ selected $.Form.Roaster .ID }}>{{ .Name }}</option>
                         {{ end}}
                     </select>
@@ -94,7 +94,7 @@
 
                 {{ template "flavours-input" (map
                     "existing" .Form.Flavours
-                    "flavours" .Data.Flavours
+                    "flavours" .Flavours
                 ) }}
 
                 <textarea name="notes" class="textarea w-full" placeholder="Notes...">{{
@@ -109,7 +109,7 @@
 </div>
 
 <h2>Coffees</h2>
-{{ range .Data.Coffees }}
+{{ range .Coffees }}
     {{ template "coffee-card" . }}
 {{ else }}
     <div class="alert alert-notice">No coffees to display</div>
